Moved contexts to their own folder
