+++ app/pathogen/arbovirus/dashboard/(visualizations)/estimate-count-by-un-region-and-arbovirus-graph.tsx
@@ -3,7 +3,7 @@ import { convertArboSFtoArbo } from "./recharts";
 import { getLabelForUNRegion } from "@/lib/un-regions";
 import { LegendConfiguration, StackedBarChart } from "../../../../../components/customs/visualizations/stacked-bar-chart";
 import { barColoursForArboviruses, sortArboviruses } from "./rechart-utils";
-import { ArboContext } from "@/contexts/pathogen-context/pathogen-contexts/arbo-context";
+import { ArboContext } from "@/contexts/pathogen-context/pathogen-contexts/arbovirus/arbo-context";
 
 interface EstimateCountByUnRegionAndArbovirusGraphProps {
   legendConfiguration: LegendConfiguration;
