+++ app/pathogen/arbovirus/layout.tsx
@@ -4,7 +4,7 @@ import getQueryClient from "@/components/customs/getQueryClient";
 import { request } from 'graphql-request';
 import { arbovirusEstimatesQuery } from "@/hooks/arbovirus/useArboData";
 import { arbovirusFiltersQuery } from "@/hooks/arbovirus/useArboFilters";
-import { ArboProviders } from "@/contexts/pathogen-context/pathogen-contexts/arbo-context";
+import { ArboProviders } from "@/contexts/pathogen-context/pathogen-contexts/arbovirus/arbo-context";
 import { GenericPathogenPageLayout } from "../generic-pathogen-page-layout";
 
 export default async function ArboLayout({
