+++ app/pathogen/sarscov2/dashboard/(map)/SarsCov2Map.tsx
@@ -2,7 +2,7 @@
 
 import React, { useContext } from "react";
 import { PathogenMap } from "@/components/ui/pathogen-map/pathogen-map";
-import { SarsCov2Context } from "@/contexts/pathogen-context/pathogen-contexts/sc2-context";
+import { SarsCov2Context } from "@/contexts/pathogen-context/pathogen-contexts/sarscov2/sc2-context";
 import { MapShadingLegend } from "@/app/pathogen/arbovirus/dashboard/(map)/MapShadingLegend";
 import { MapEstimateSummary } from "@/components/ui/pathogen-map/map-estimate-summary";
 import { MapSymbology } from "./map-config";
