+++ app/pathogen/sarscov2/dashboard/(map)/sars-cov-2-estimate-pop-up-content.tsx
@@ -1,6 +1,6 @@
 import { GenericMapPopUp, GenericMapPopUpWidth, HeaderConfigurationTextAlignment } from "@/components/ui/pathogen-map/map-pop-up/generic-map-pop-up";
 import { PopUpContentRowType } from "@/components/ui/pathogen-map/map-pop-up/pop-up-content-rows";
-import { SarsCov2Estimate } from "@/contexts/pathogen-context/pathogen-contexts/sc2-context"
+import { SarsCov2Estimate } from "@/contexts/pathogen-context/pathogen-contexts/sarscov2/sc2-context"
 
 interface SarsCov2EstimatePopupContentProps {
   estimate: SarsCov2Estimate;
