+++ app/pathogen/sarscov2/dashboard/(visualizations)/cumulative-number-of-serosurveys-published-over-time.tsx
@@ -1,7 +1,7 @@
 import uniq from "lodash/uniq";
 import { useMemo } from 'react';
 import { AreaChart } from "@/components/customs/visualizations/area-chart";
-import { SarsCov2Context, SarsCov2Estimate } from "@/contexts/pathogen-context/pathogen-contexts/sc2-context";
+import { SarsCov2Context, SarsCov2Estimate } from "@/contexts/pathogen-context/pathogen-contexts/sarscov2/sc2-context";
 import { generateRandomColour, generateRange, typedGroupBy, typedObjectEntries, typedObjectFromEntries } from "@/lib/utils";
 import { useContext } from "react";
 import parseISO from 'date-fns/parseISO';
