+++ app/pathogen/sarscov2/dashboard/(visualizations)/modelled-seroprevalence-by-who-region.tsx
@@ -2,7 +2,7 @@ import { useContext, useMemo } from "react";
 import parseISO from 'date-fns/parseISO';
 
 import { LegendConfiguration } from "@/components/customs/visualizations/stacked-bar-chart";
-import { SarsCov2Context, SarsCov2Estimate } from "@/contexts/pathogen-context/pathogen-contexts/sc2-context";
+import { SarsCov2Context, SarsCov2Estimate } from "@/contexts/pathogen-context/pathogen-contexts/sarscov2/sc2-context";
 import { dateToMonthCount, monthCountToMonthYearString, monthYearStringToMonthCount } from "@/lib/time-utils";
 import { WhoRegion } from "@/gql/graphql";
 import { BestFitCurveLineChart } from "@/components/customs/visualizations/best-fit-curve-line-chart";
