+++ app/pathogen/sarscov2/dashboard/filters.tsx
@@ -3,7 +3,7 @@
 import React, { useContext } from "react";
 import { Filters } from "@/components/customs/filters";
 import { FilterableField } from "@/components/customs/filters/available-filters";
-import { SarsCov2Context } from "@/contexts/pathogen-context/pathogen-contexts/sc2-context";
+import { SarsCov2Context } from "@/contexts/pathogen-context/pathogen-contexts/sarscov2/sc2-context";
 import { useSarsCov2Filters } from "@/hooks/sarscov2/useSarsCov2Filters";
 
 interface SarsCov2FiltersProps {
